Make previous commits python2 compatible
Use six.raise_from instead of its python3 variant.

Replace reportlib_resources useage with inspect.
The reportlib_resources functionality was not available for one that is
also compatible with python2.
Also some older versions of reportlib_resources fail when used this way.

Remove arguent use_fallback for xmlschema as it was only introduced in
v1.2.3.